# Refactor String#char_bytesize_at