I tried to wrap my mind around how `String#char_bytesize_at` works and ended up refactoring it to this version to facilitate understanding. I used [the UTF-8 bit patterns](https://en.wikipedia.org/wiki/UTF-8#Description) to make it clearer what’s happening in this method.

Note: I’m not sure why the current implementation uses the `0x90`, `0xc2` and `0xf4` values for comparisons, so my assumption that this method counts the bytesize of an UTF-8 sequence might be wrong.