@chastell 2018-01-01
Right, I actually had this thought right after opening this PR. 🤦‍♀️ 

A quick bench to the tune of

```crystal
string = String.new(Bytes[104, 101, 108, 108, 111, 32, 255, 32, 255, 32, 119, 111, 114, 108, 100, 33])

Benchmark.ips do |bench|
  bench.report("old") { (0...16).map { |index| string.old(index) } }
  bench.report("new") { (0...16).map { |index| string.new(index) } }
end
```

shows a 6% slowdown:

```
old   8.85M (112.95ns) (± 3.25%)       fastest
new   8.37M (119.48ns) (± 2.30%)  1.06× slower
```

I’ll work on this PR to make it faster, but in the meantime: how can I best create a plausible set of representative strings for benching?