@chastell 2018-01-01
Interestingly, on a random string of bytes the new implementation is actually significantly faster, which strongly suggests I’m missing something here:

```crystal
string = String.new(Random.new.random_bytes(256))

Benchmark.ips do |bench|
  bench.report("old") { string.size.times { |index| string.old(index) } }
  bench.report("new") { string.size.times { |index| string.new(index) } }
end
```

```
old 345.98M (  2.89ns) (± 3.16%)  1.40× slower
new 484.79M (  2.06ns) (± 2.10%)       fastest

old  349.6M (  2.86ns) (± 2.75%)  1.38× slower
new 481.84M (  2.08ns) (± 2.84%)       fastest

old  347.7M (  2.88ns) (± 3.09%)  1.38× slower
new  479.3M (  2.09ns) (± 4.00%)       fastest
```

🤔