@straight-shoota 2018-01-01
The individual bachmark measurements are so small (only nanoseconds!) that there won't be any reliable results. You could wrap the individual examples in `1000.times { ... }` to get a more feasible number.
And it would be great to have a benchmark covering consecutive calls through `each_byte_index_and_char_index`, like `String#size` (when it's not already initialized).

Did you look at other implementations? No need to reinvent every wheel. `Char::Reader#decode_char_at` is also very similar, allthough it works a bit different. I guess it makes sense to keep both implemented differently but improvements to the algorithm could be added there as well.