@chastell 2018-01-01
Ah, yes, the first `String#size` call in my benchmark precomputes the size. 🤦‍♂️ 🤦‍♂️ 🤦‍♂️ 

When I removed it, a random 256-byte string takes exactly the same amount of time:

```crystal
string = String.new(Random.new.random_bytes(256))

Benchmark.ips do |bench|
  bench.report("old") { 256.times { |index| string.old(index) } }
  bench.report("new") { 256.times { |index| string.new(index) } }
end
```

```
old 552.94M (  1.81ns) (± 2.21%)  1.00× slower
new 554.01M (  1.81ns) (± 1.99%)       fastest

old 555.95M (   1.8ns) (± 2.39%)       fastest
new 555.41M (   1.8ns) (± 2.35%)  1.00× slower

old 562.24M (  1.78ns) (± 2.58%)  1.00× slower
new  562.4M (  1.78ns) (± 2.33%)       fastest
```

> The individual bachmark measurements are so small (only nanoseconds!) that there won't be any reliable results. You could wrap the individual examples in `1000.times { ... }` to get a more feasible number.

Hmmm, I thought this is exactly what `Benchmark.ips` does, no? It even shows the standard deviation, so we know how reliable they are.

I’ll look at your other pointers, thanks so much!