@chastell 2018-01-01
OK, I think the culprit here might be that the whole block is compiled out as a no-op. 🤦‍♂️ 

(This is a very teaching start to the new year, thanks for bearing with me…)