@chastell 2018-01-01
OK, just for posterity:

```crystal
string = String.new(Random.new.random_bytes(256))

Benchmark.ips do |bench|
  bench.report("old") { (0...256).map { |index| string.old(index) } }
  bench.report("new") { (0...256).map { |index| string.new(index) } }
end
```

```
old  933.0k (  1.07µs) (± 2.65%)  1.16× slower
new   1.08M (927.91ns) (± 6.22%)       fastest

old 938.74k (  1.07µs) (± 2.42%)       fastest
new 908.57k (   1.1µs) (± 2.02%)  1.03× slower

old 906.63k (   1.1µs) (± 2.65%)       fastest
new 883.05k (  1.13µs) (± 2.42%)  1.03× slower

old  934.1k (  1.07µs) (± 2.34%)       fastest
new 922.33k (  1.08µs) (± 2.30%)  1.01× slower

old 914.57k (  1.09µs) (± 2.56%)  1.17× slower
new   1.07M (931.93ns) (± 5.93%)       fastest
```

Not bad. 😉 

I’m still not sure why the current implementation uses byte values like `0x90`, `0xc2` and `0xf4`, though. 🤷‍♂️