@straight-shoota 2018-01-01
`Benchmark.ips` runs each example multiple times but each is measured on it's own and these measurements are used to calculate average and deviation. When each measurement is too small of a time span (say less than a few microseconds) the results will not be reliable.

Unfortunately, the results from `#size` are not useful because it is cached and and the cache is already populated in the constructor. You could add a non-caching implementation of the size method to properly test it.