@straight-shoota 2018-01-01
Oh sry, I mixed that up. Yes, this is only true for `Benchmark.bm`.

I guess I'm doing to many things simultaneously... no real multithreading support.