@straight-shoota 2018-01-01
@RX14 The 1.6 ns is just reading `@length`