@RX14 2018-01-01
We need a JMH-style blackhole class to get accurate benchmark results.