@chastell 2018-01-01
OK, I reimplemented it _for speed_™ and the results for a random 10000-byte string are quite optimistic:

```
old  19.92k ( 50.19µs) (± 1.97%)  1.01× slower
new   20.1k ( 49.75µs) (± 1.57%)       fastest

old  20.25k ( 49.38µs) (± 1.05%)       fastest
new  20.14k ( 49.65µs) (± 1.38%)  1.01× slower

old  20.18k ( 49.56µs) (± 0.69%)  1.00× slower
new  20.25k ( 49.39µs) (± 0.97%)       fastest
```

Thanks so much for the pointer to `Char::Reader#decode_char_at`, that explains the shape of the current implemenation! I’m still not sure why `0xc1` is a valid byte sequence there, though. 🤷‍♂️