@straight-shoota 2018-01-01
I'd assume `0xC2` might just have been a typo or some other mistake (off by one). It's very unlikely to cause any issues therefore it was not noticed.

I think you can safely skip the `0b0_0000000 <= byte` comparisons, thats the minimum for `UInt8` anyway. But I like the `subsequent_utf8_byte_at?` method, makes it easier readable and shouldn't add any costs.